Create new basic HTML structure for all sections
Since the page will be styled using flexbox, it is much easier to
have each section have some defined content (div) and expand to
fill the viewport. This is much easier than styling each section
at a time because the other sections (div) because a full
foundation makes adding and styling elements easier than without it.