Replace section four content div with container div
This naming convention makes more sense for the purpose of the
project. All the content in each of the sections have been nested
inside a 'container' and have been aligned as flex items. Therefore,
this naming convention was changed for all the sections, with each
section having exactly one div with the class name 'container' to
contain the contents of the page.