Change body height to a defined pixel amount
Depending on the size of a user's device, elements may appear
to be further apart than they actually are if the body's height
is measured in terms of the visible viewport (essentially, the
elements will look more spread apart if you zoom out while viewing
on a computer). Thus, a defined pixel amount that does not change
as the viewport gets larger or smaller solves the issue.